Move updatePointSize into engine/src/pointSizeScaling.ts
Joins the existing screen-space point-size scaling logic in the same
module since both operate on the same per-file point-size state.
